reset bitmaps of dyn bundles containing ptr slots
this could sneak an allocation from a child instance into
its parent instance, which would then lead into a use-after-free
after the child and its allocator are disposed.
this makes fresh copies of bitmaps of bundles that both are dynamic
and contain a pointer value.
note that static bundles are fine, because the instance
is copied anyway, and dynamic bundles not containing pointers
are fine because there is nothing to leak.
also, only the bitmaps need to be copied, since even if the
child uses the parent's buffer, the parent will reallocate
that value if the bit is not set in the bitmap.