pin allocs to control instructions
consider the following snippets:

(1) new = alloc(...)
    loop:
      use new

(2) loop:
      new = alloc(...)
      use new

they are obviously semantically different, but previously
the ir had no way to distinguish between them, since if
`alloc` was pure, they would be equivalent and global code
motion is allowed to choose either.

you could force gcm to emit (1) by using a suitable chain
of MOVFs, ie. something like
  0001 ptr ALLOC ...
  0002 fx  CALL ... 0001 ...
  0003 ptr MOVF 0001 0002
but this is not quite correct either, as gcm is allowed to
duplicate 0001 (though i'm not sure if there's a way to trigger
this in practice).

this commit fixes the problem by pinning the alloc to a control
instruction similar to how PHIs are pinned. this makes the IR
for snippets (1) and (2) different, and is also similar to what
is done in eg. the `Simple` tutorial.