bandaid fix for schedule assert fail
the algorithm that computes the cmat is subtly wrong
here in the case where A dominates B and B is a loop.
the algorithm doesn't correctly propagate cmat(A) to
cmat(B) because the loop body is a precedessor.

the current algorithm should be scrapped and replaced
with a union over dominators and postdominators (i think?)